requirements.txt
websockets==12.0
httpx==0.27.0
python-dotenv==1.1.0
pydantic==1.10.14
